requirements.txt
appdirs==1.4.4
arrow==1.2.2
attrs==20.3.0
boto3==1.17.26
botocore==1.20.26
bokeh==2.4.3
emmett==2.2.1
jinja2
matplotlib
numpy
pipenv==2020.11.15
plotly==5.9.0
plotly-express==0.4.1
pyarrow==8.0.0
python-dateutil==2.8.2
python-dotenv
PyYAML
requests==2.28.1
streamlit==1.11.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.10
supabase
git+https://github.com/samdobson/streamlit-toggle
websockets
virtualenv==20.4.4
virtualenv-clone==0.5.4
cfbd
streamlit
altair==4.2.0
awesome-panel==20200512.1
pandas
timedelta